necessary fix for replace the host when ip is 0.0.0.0 when opening a browser from tray icon
