+++ src/main/java/runwar/Server.java
@@ -60,6 +60,7 @@ import static io.undertow.servlet.Servlets.defaultContainer;
 import static io.undertow.servlet.Servlets.deployment;
 import static runwar.logging.RunwarLogger.CONTEXT_LOG;
 import static runwar.logging.RunwarLogger.LOG;
+import runwar.util.Utils;
 
 public class Server {
 
@@ -175,7 +176,7 @@ public class Server {
     }
 
     private synchronized void requisitionPorts() {
-        LOG.debug("HOST to be bound:"+serverOptions.host());
+        LOG.debug("HOST to be bound:" + serverOptions.host());
         ports = new PortRequisitioner(serverOptions.host());
         ports.add("http", serverOptions.httpPort());
         ports.add("stop", serverOptions.stopPort());
@@ -239,12 +240,12 @@ public class Server {
         Builder serverBuilder = Undertow.builder();
         setUndertowOptions(serverBuilder);
 
-        LOG.debug("SERVER BUILDER:"+serverOptions.httpEnable());
+        LOG.debug("SERVER BUILDER:" + serverOptions.httpEnable());
         if (serverOptions.httpEnable()) {
-            LOG.debug("Server Builder - PORT:"+ports.get("http").socket+" HOST:"+host);
+            LOG.debug("Server Builder - PORT:" + ports.get("http").socket + " HOST:" + host);
             serverBuilder.addHttpListener(ports.get("http").socket, host);
-        }else{
-        LOG.info("HTTP Enabled:"+serverOptions.httpEnable());
+        } else {
+            LOG.info("HTTP Enabled:" + serverOptions.httpEnable());
         }
 
         
@@ -256,8 +257,8 @@ public class Server {
             }
             serverBuilder.setServerOption(UndertowOptions.ENABLE_HTTP2, true);
             //serverBuilder.setSocketOption(Options.REUSE_ADDRESSES, true);
-        }else{
-            LOG.info("HTTP2 Enabled:"+serverOptions.http2Enable());
+        } else {
+            LOG.info("HTTP2 Enabled:" + serverOptions.http2Enable());
         }
 
         if (serverOptions.sslEnable()) {
@@ -290,8 +291,8 @@ public class Server {
                 e.printStackTrace();
                 System.exit(1);
             }
-        }else{
-            LOG.info("HTTP sslEnable:"+serverOptions.sslEnable());
+        } else {
+            LOG.info("HTTP sslEnable:" + serverOptions.sslEnable());
         }
 
         if (serverOptions.ajpEnable()) {
@@ -301,8 +302,8 @@ public class Server {
                 // if no options is set, default to the large packet size
                 serverBuilder.setServerOption(UndertowOptions.MAX_AJP_PACKET_SIZE, 65536);
             }
-        }else{
-            LOG.info("HTTP ajpEnable:"+serverOptions.ajpEnable());
+        } else {
+            LOG.info("HTTP ajpEnable:" + serverOptions.ajpEnable());
         }
 
         securityManager = new SecurityManager();
@@ -327,9 +328,9 @@ public class Server {
                 serverOptions.contentDirs(warFile.getAbsolutePath());
             }
             serverOptions.warFile(warFile);
-        }else{
-            LOG.info("HTTP warFile exists:"+warFile.exists());
-            LOG.info("HTTP warFile isDirectory:"+warFile.isDirectory());
+        } else {
+            LOG.info("HTTP warFile exists:" + warFile.exists());
+            LOG.info("HTTP warFile isDirectory:" + warFile.isDirectory());
         }
         if (!warFile.exists()) {
             throw new RuntimeException("war does not exist: " + warFile.getAbsolutePath());
@@ -343,8 +344,8 @@ public class Server {
             // just in case
             Thread.sleep(200);
             System.exit(0);
-        }else{
-            LOG.info("HTTP background:"+serverOptions.background());
+        } else {
+            LOG.info("HTTP background:" + serverOptions.background());
         }
 
         File webinf = serverOptions.webInfDir();
@@ -594,7 +595,7 @@ public class Server {
                     });
                 }
 
-                if (serverOptions.debug() && serverOptions.testing() &&exchange.getRequestPath().endsWith("/dumprunwarrequest")) {
+                if (serverOptions.debug() && serverOptions.testing() && exchange.getRequestPath().endsWith("/dumprunwarrequest")) {
                     new RequestDumper().handleRequest(exchange);
                 } else {
                     super.handleRequest(exchange);
@@ -1140,7 +1141,7 @@ public class Server {
                 openbrowserURL = protocol + "://" + host + ":" + portNumber + openbrowserURL;
             }
             // if binding to all IPs, swap out with localhost.
-            openbrowserURL = replaceHost(openbrowserURL, "0.0.0.0", "127.0.0.1");
+            openbrowserURL = Utils.replaceHost(openbrowserURL, "0.0.0.0", "127.0.0.1");
 
             LOG.info("Waiting up to " + (timeout / 1000) + " seconds for " + host + ":" + portNumber + "...");
             try {
@@ -1157,21 +1158,6 @@ public class Server {
         }
     }
 
-    public static String replaceHost(String openbrowserURL, String oldHost, String newHost) {
-        String url = openbrowserURL;
-        try {
-            URL address = new URL(openbrowserURL);
-            String host = address.getHost();
-            if (host.equalsIgnoreCase(oldHost)) {
-                    URL ob=new URL(address.getProtocol(), newHost, address.getPort() , address.getFile());
-                    openbrowserURL=ob.toString();
-            }
-        } catch (MalformedURLException ex) {
-            ex.printStackTrace();
-            openbrowserURL = url;
-        }
-        return openbrowserURL;
-    }
 
     public ServerOptions getServerOptions() {
         return serverOptions;
